team8/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
transformers>=4.30.0
huggingface-hub>=0.17.0
python-dotenv>=1.0.0
